Move hydrate to separate file. Add comments.
