+++ src/lib/parts.ts
@@ -274,9 +274,9 @@ export class NodePart implements Part {
           new TemplateInstance(template, value.processor, this.options);
 
       // This is the marker content that would have been generated by this
-      // TemplateResult.
-      const markerWithHash = `lit-part ${value.digest}`;
-      if (this.startNode.textContent === markerWithHash) {
+      // TemplateResult's template during SSR.
+      const markerWithDigest = `lit-part ${value.digest}`;
+      if (this.options.hydrate && this.startNode.textContent === markerWithDigest) {
         // We got a match, so we're rehydrating DOM pre-rendered with this
         // TemplateResult's Template.
 
@@ -297,9 +297,13 @@ export class NodePart implements Part {
         const fragment = instance._clone();
         instance.update(value.values);
         this.__commitNode(fragment);
-        // if (ssr) {
-        this.startNode.textContent = markerWithHash;
-        // }
+        // In "SSR" mode we always render two markers and write the template
+        // digest to the start marker so that we can compare during hydration.
+        // This is only for testing, and only for now, since real SSR is done
+        // with a different server implementation.
+        if (this.options.ssr) {
+          this.startNode.textContent = markerWithDigest;
+        }
       }
       this.value = instance;
     }
