+++ src/lib/render.ts
@@ -18,7 +18,7 @@
 
 import {removeNodes} from './dom.js';
 import {NodePart} from './parts.js';
-import {PartInfo, RenderOptions} from './render-options.js';
+import {RenderOptions} from './render-options.js';
 import {templateFactory} from './template-factory.js';
 
 export const parts = new WeakMap<Node, NodePart>();
@@ -56,77 +56,3 @@ export const render =
       part.commit();
     };
 
-export const hydrate =
-    (result: unknown,
-     container: Element|DocumentFragment,
-     options?: Partial<RenderOptions>) => {
-      console.log('hydrate', container.childNodes[0].textContent);
-
-      let part = parts.get(container);
-
-      if (part === undefined) {
-        let rootPart: PartInfo|undefined = undefined;
-        const partStack: PartInfo[] = [];
-        const walker = document.createTreeWalker(
-            container, NodeFilter.SHOW_COMMENT, null, false);
-        let node: Comment|null;
-        while ((node = walker.nextNode() as Comment | null) !== null) {
-          if (node.nodeType === Node.COMMENT_NODE) {
-            if (node.textContent!.startsWith('lit-part')) {
-              // This is an NodePart opening marker
-              const partInfo = {
-                startNode: node,
-                endNode: undefined as unknown as Node,
-              };
-              if (partStack.length > 0) {
-                const parentInfo = partStack[partStack.length - 1];
-                if (parentInfo.children === undefined) {
-                  parentInfo.children = [];
-                }
-                parentInfo.children.push(partInfo);
-              } else {
-                console.assert(
-                    rootPart === undefined,
-                    'there should be exactly one root part in a render container');
-                rootPart = partInfo;
-              }
-              partStack.push(partInfo);
-            } else if (node.textContent!.startsWith('/lit-part')) {
-              // This is an NodePart closing marker
-              const partInfo = partStack.pop()!;
-              partInfo.endNode = node;
-            }
-          }
-        }
-        console.assert(
-            rootPart !== undefined,
-            'there should be exactly one root part in a render container');
-
-        // If not specified, assume template data has changed since pre-rendering.
-        // Assuming data has changed may cause unnecessary DOM reconstruction;
-        // however, this is better than displaying the wrong data.
-        if (options === undefined) {
-          options = { dataChanged: true };
-        } else if (options.dataChanged === undefined) {
-          options.dataChanged = true;
-        }
-
-        part = new NodePart({
-          templateFactory,
-          ...options,
-          prerenderedParts: rootPart!.children
-        });
-        part.startNode = rootPart!.startNode;
-        part.endNode = rootPart!.endNode;
-        parts.set(container, part);
-      } else {
-        part = new NodePart({
-          templateFactory,
-          ...options,
-        });
-        parts.set(container, part);
-        part.appendInto(container);
-      }
-      part.setValue(result);
-      part.commit();
-    };
